lord-merick-farrow: Moved cypress posts to constant file
