+++ cypress/integration/homepage.spec..js
@@ -1,4 +1,5 @@
 /// <reference types="Cypress" />
+import posts from '../constants/posts'
 
 describe('Given the homepage', () => {
   beforeEach(() => {
@@ -10,75 +11,6 @@ describe('Given the homepage', () => {
   })
 
   it('should display all posts', () => {
-    const posts = [
-      {
-        title: 'Lord Merick Farrow',
-        origin: 'Descent: Journeys in the Dark - Second Edition',
-        date: '24 Jan 2021',
-        href: 'lord-merick-farrow',
-      },
-      {
-        title: 'Elder Mok',
-        origin: 'Descent: Journeys in the Dark - Second Edition',
-        date: '16 Jan 2021',
-        href: 'elder-mok',
-      },
-      {
-        title: 'Karnon',
-        origin: 'Descent: Journeys in the Dark - Second Edition',
-        date: '10 Jan 2021',
-        href: 'karnon',
-      },
-      {
-        title: 'Valyndra',
-        origin: 'Descent: Journeys in the Dark - Second Edition',
-        date: '29 Dec 2020',
-        href: 'valyndra',
-      },
-      {
-        title: 'Cave Spiders',
-        origin: 'Descent: Journeys in the Dark - Second Edition',
-        date: '24 Dec 2020',
-        href: 'cave-spiders',
-      },
-      {
-        title: 'Striders',
-        origin: 'Horizon Zero Dawn - The Board Game',
-        date: '20 Dec 2020',
-        href: 'striders',
-      },
-      {
-        title: 'Belthir',
-        origin: 'Descent: Journeys in the Dark - Second Edition',
-        date: '15 Dec 2020',
-        href: 'belthir',
-      },
-      {
-        title: 'Watchers',
-        origin: 'Horizon Zero Dawn - The Board Game',
-        date: '10 Dec 2020',
-        href: 'watchers',
-      },
-      {
-        title: 'Campaign Battle - Belthir & Sorcerers',
-        origin: 'Descent: Journeys in the Dark - Second Edition',
-        date: '06 Dec 2020',
-        href: 'campaign-battle-01',
-      },
-      {
-        title: 'Sorcerers',
-        origin: 'Descent: Journeys in the Dark - Second Edition',
-        date: '04 Dec 2020',
-        href: 'sorcerers',
-      },
-      {
-        title: 'First Post',
-        origin: 'Beginnings',
-        date: '03 Dec 2020',
-        href: 'first-post',
-      },
-    ]
-
     cy.findAllByTestId('postListItem')
       .should('have.length', posts.length)
       .each((item, index) => {
